+++ app/src/components/grids/collectionsGrid.tsx
@@ -8,7 +8,7 @@ import CollectionDataType from "../../types/CollectionDataType";
 import React, { useContext, useRef, useState } from "react";
 import { totalNumPages } from "../../utils/utils";
 import DCSimpleGrid from "../dcSimpleGrid/dcSimpleGrid";
-import DCCard from "../cards/DCCard";
+import DCCard from "../dcCard/DCCard";
 import { useTooltipOffset } from "@/src/hooks/useTooltipOffset";
 
 export const CollectionsGrid = ({ data }: any) => {
